+++ R/ts_makeframes.R
@@ -9,6 +9,7 @@
 #' @param l_indices (Optional) numeric, a vector of layer indices specifying which layers are to be plotted. Should contain 3 values for an RGB image or a single value for a discrete or gradient image. By default, chooses the first three layers if \code{r_type} is \code{"RGB"} and the first layer if \code{r_type} is \code{"discrete"} or \code{"gradient"}.
 #' @param hillshade (Optional) A raster layer. If one is provided, it will be used as the base layer and plotted with \link[RStoolbox]{ggR} while the raster layers from x_list will be plotted over it with a default \code{alpha} of 0.5. By default, no hillshade is used.
 #' @param alpha (Optional) numeric. The opacity of the plot. Default is 1.
+#' @param ... (Optional) further arguments for the plotting.
 #' @return A list of ggplots
 #' @details A linear percent stretch will be applied to each band to improve contrast.
 #' #\deqn{(X - Low_{in}) * \frac{((High_{out}-Low_{out})}{(High_{in}-Low_{in}))}  + Low_{out}}
@@ -44,7 +45,7 @@
 #' # as from the desired layers
 #'r_frames <- ts_makeframes(x_list = r_list_out,samplesize = 10,
 #'                           l_indices = c(1,4,3))
-ts_makeframes <- function(x_list,r_type = NULL,minq = 0.02,maxq = 0.98,samplesize = 1000,blacken_NA=FALSE,l_indices=NULL,alpha=NULL,hillshade=NULL){
+ts_makeframes <- function(x_list,r_type = NULL,minq = 0.02,maxq = 0.98,samplesize = 1000,blacken_NA=FALSE,l_indices=NULL,alpha=NULL,hillshade=NULL,...){
   
   #If r_type not provided, guess from the first raster object in the list
   if(is.null(r_type)){
@@ -80,10 +81,18 @@ ts_makeframes <- function(x_list,r_type = NULL,minq = 0.02,maxq = 0.98,samplesiz
     if(is.null(alpha)){
       alpha=0.5
     }
+    
+    #transform if necessary
+    if(!compareCRS(hillshade,x_list[[1]])){
+      print("Reprojecting Hillshade")
+      hillshade <- projectRaster(from = hillshade,(x_list[[1]]))
+    }
+    hillshade <- crop(hillshade,x_list[[1]])
+    
     # make a hillshade annotation layer
-    hillshade_layer <- RStoolbox::ggR(projectRaster(hillshade,x_list[[1]]),ggLayer = TRUE)
+    hillshade_layer <- RStoolbox::ggR(hillshade,ggLayer = TRUE)
     # pass it to ggbmap
-    r_ggplots <- .ts_makeframes(x_list = r_list_out_stretched,r_type = r_type,gglayer=TRUE,alpha=alpha,hillshade_layer=hillshade_layer)
+    r_ggplots <- .ts_makeframes(x_list = r_list_out_stretched,r_type = r_type,gglayer=TRUE,alpha=alpha,hillshade_layer=hillshade_layer,...)
     #plot the layers over the hillshade
     #r_ggplots <- lapply(r_ggplots,FUN = function(x){
     #  hillshade_layer+x
@@ -93,7 +102,7 @@ ts_makeframes <- function(x_list,r_type = NULL,minq = 0.02,maxq = 0.98,samplesiz
       alpha=1
     }
     #make the plots
-    r_ggplots <- .ts_makeframes(x_list = r_list_out_stretched,r_type = r_type,gglayer=FALSE,alpha=alpha)
+    r_ggplots <- .ts_makeframes(x_list = r_list_out_stretched,r_type = r_type,gglayer=FALSE,alpha=alpha,...)
   }
   
   r_ggplots <- .ts_set_frametimes(r_ggplots , .ts_get_frametimes(x_list))