+++ NEWS.md
@@ -1,7 +1,11 @@
 ***
 
-## rtsVis 0.0.2
-Various Bug fixes.
+## rtsVis 0.0.3
+Added additional plot functions to ts_flow_frames.
+
+Improved hillshade support for ts_makeframes.
+
+Various bugfixes.